+++ datar/dplyr/relocate.py
@@ -15,7 +15,7 @@ def relocate(
         *args: Any,
         _before: Optional[Union[int, str]] = None,
         _after: Optional[Union[int, str]] = None,
-        _base0: Optional[bool] = None,
+        base0_: Optional[bool] = None,
         **kwargs: Any
 ) -> DataFrame:
     """change column positions
@@ -30,7 +30,7 @@ def relocate(
         _before: and
         _after: Destination. Supplying neither will move columns to
             the left-hand side; specifying both is an error.
-        _base0: Whether `_before` and `_after` are 0-based if given by indexes.
+        base0_: Whether `_before` and `_after` are 0-based if given by indexes.
             If not provided, will use `datar.base.get_option('index.base.0')`
 
     Returns:
@@ -48,7 +48,7 @@ def relocate(
         all_columns,
         *args,
         **kwargs,
-        _base0=_base0,
+        base0_=base0_,
         _group_vars=gvars
     )
     if _before is not None and _after is not None:
@@ -59,14 +59,14 @@ def relocate(
     # length = len(all_columns)
     if _before is not None:
         where = min(_eval_select(
-            all_columns, _before, _group_vars=[], _base0=_base0
+            all_columns, _before, _group_vars=[], base0_=base0_
         )[0])
         if where not in to_move:
             to_move.append(where)
 
     elif _after is not None:
         where = max(_eval_select(
-            all_columns, _after, _group_vars=[], _base0=_base0
+            all_columns, _after, _group_vars=[], base0_=base0_
         )[0])
         if where not in to_move:
             to_move.insert(0, where)