+++ docs/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.3.0
+- Adopt `pipda` v0.3.0
+
+Breaking changes:
+
+- Rename argument `dtypes` of `unchop` and `unnest` back to `ptype`
+- Change all `_base0` to `base0_`
+- Change argument `how` of `tidyr.drop_na` to `how_`
+
 ## 0.2.3
 - Fix compatibility with `pandas` `v1.2.0~4` (#20, thanks to @antonio-yu)
 - Fix base.table when inputs are factors and exclude is NA;
@@ -63,7 +72,7 @@ Changed:
 ## 0.0.6
 - Add `options`, `get_option` and `options_context` to `datar.base` to allow set/get global options
 - Add options: `dplyr.summarise.inform`
-- Add `_base0` argument to all related APIs
+- Add `base0_` argument to all related APIs
 - Add `nycflights13` datasets
 - Support slice_head/slice_tail for grouped data